lab4/pyproject.toml
[tool.poetry]
name = "lab4"
version = "0.1.0"
description = ""
authors = ["Nikita Trynus <trynuspoc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
poetry = "^1.2.2"
typer = "^0.7.0"
pycryptodome = "^3.16.0"
black = "^22.10.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
